That value needs to be a Long, not an Int

+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/verify/VerifiableLocation.scala
@@ -5,5 +5,5 @@ import java.net.URI
 case class VerifiableLocation(
   uri: URI,
   checksum: Checksum,
-  length: Option[Int]
+  length: Option[Long]
 )
